refactor: extract 7 sub-components from classroom-modal.jsx
Extract ErrorDisplay, ClassCodeDisplay, GoogleCourseList, TeacherClassDetail,
TeacherCreateForm, TeacherPostAssignment, and StudentJoinForm into individual
files under components/classroom-modal/. These shared sub-components can be
imported by both the student modal and the teacher fullscreen modal.

Reduces classroom-modal.jsx from ~2100 lines to ~920 lines with no
behavioral changes.
